+++ sklearn/ensemble/gradient_boosting.py
@@ -35,7 +35,7 @@ from .base import BaseEnsemble
 from ..base import BaseEstimator
 from ..base import ClassifierMixin
 from ..base import RegressorMixin
-from ..utils import check_random_state, array2d, check_arrays, column_or_1d
+from ..utils import check_random_state, check_array, check_X_y, column_or_1d
 from ..utils.extmath import logsumexp
 from ..externals import six
 from ..feature_selection.from_model import _LearntSelectorMixin
@@ -751,8 +751,7 @@ class BaseGradientBoosting(six.with_metaclass(ABCMeta, BaseEnsemble,
             self._clear_state()
 
         # Check input
-        X, = check_arrays(X, dtype=DTYPE, sparse_format="dense")
-        y = column_or_1d(y, warn=True)
+        X, y = check_X_y(X, y, dtype=DTYPE)
         n_samples, n_features = X.shape
         self.n_features = n_features
         random_state = check_random_state(self.random_state)
@@ -778,7 +777,7 @@ class BaseGradientBoosting(six.with_metaclass(ABCMeta, BaseEnsemble,
                                  % (self.n_estimators,
                                     self.estimators_.shape[0]))
             begin_at_stage = self.estimators_.shape[0]
-            y_pred = self.decision_function(X)
+            y_pred = self._decision_function(X)
             self._resize_state()
 
         # fit the boosting stages
@@ -872,6 +871,13 @@ class BaseGradientBoosting(six.with_metaclass(ABCMeta, BaseEnsemble,
         score = self.init_.predict(X).astype(np.float64)
         return score
 
+    def _decision_function(self, X):
+        # for use in inner loop, not raveling the output in single-class case,
+        # not doing input validation.
+        score = self._init_decision_function(X)
+        predict_stages(self.estimators_, X, self.learning_rate, score)
+        return score
+
     def decision_function(self, X):
         """Compute the decision function of ``X``.
 
@@ -882,15 +888,16 @@ class BaseGradientBoosting(six.with_metaclass(ABCMeta, BaseEnsemble,
 
         Returns
         -------
-        score : array, shape = [n_samples, k]
+        score : array, shape = [n_samples, n_classes] or [n_samples]
             The decision function of the input samples. The order of the
             classes corresponds to that in the attribute `classes_`.
-            Regression and binary classification are special cases with
-            ``k == 1``, otherwise ``k==n_classes``.
+            Regression and binary classification produce an array of shape
+            [n_samples].
         """
-        X = array2d(X, dtype=DTYPE, order="C")
-        score = self._init_decision_function(X)
-        predict_stages(self.estimators_, X, self.learning_rate, score)
+        X = check_array(X, dtype=DTYPE, order="C")
+        score = self._decision_function(X)
+        if score.shape[1] == 1:
+            return score.ravel()
         return score
 
     def staged_decision_function(self, X):
@@ -912,7 +919,7 @@ class BaseGradientBoosting(six.with_metaclass(ABCMeta, BaseEnsemble,
             Regression and binary classification are special cases with
             ``k == 1``, otherwise ``k==n_classes``.
         """
-        X = array2d(X, dtype=DTYPE, order="C")
+        X = check_array(X, dtype=DTYPE, order="C")
         score = self._init_decision_function(X)
         for i in range(self.estimators_.shape[0]):
             predict_stage(self.estimators_, i, X, self.learning_rate, score)