+++ sklearn/preprocessing/label.py
@@ -20,7 +20,7 @@ from ..utils.fixes import np_version
 from ..utils.fixes import sparse_min_max
 from ..utils.fixes import astype
 from ..utils import deprecated, column_or_1d
-from ..utils.validation import check_arrays
+from ..utils.validation import check_array
 from ..utils.multiclass import unique_labels
 from ..utils.multiclass import type_of_target
 
@@ -453,7 +453,7 @@ def label_binarize(y, classes, neg_label=0, pos_label=1,
     """
     if not isinstance(y, list):
         # XXX Workaround that will be removed when list of list format is dropped
-        y, = check_arrays(y)
+        y = check_array(y, accept_sparse='csr', ensure_2d=False)
     if neg_label >= pos_label:
         raise ValueError("neg_label={0} must be strictly less than "
                          "pos_label={1}.".format(neg_label, pos_label))