+++ sklearn/svm/tests/test_svm.py
@@ -294,11 +294,11 @@ def test_decision_function():
     clf.fit(X, Y)
     dec = np.dot(X, clf.coef_.T) + clf.intercept_
     prediction = clf.predict(X)
-    assert_array_almost_equal(dec, clf.decision_function(X))
+    assert_array_almost_equal(dec.ravel(), clf.decision_function(X))
     assert_array_almost_equal(
         prediction,
-        clf.classes_[(clf.decision_function(X) > 0).astype(np.int).ravel()])
-    expected = np.array([[-1.], [-0.66], [-1.], [0.66], [1.], [1.]])
+        clf.classes_[(clf.decision_function(X) > 0).astype(np.int)])
+    expected = np.array([-1., -0.66, -1., 0.66, 1., 1.])
     assert_array_almost_equal(clf.decision_function(X), expected, 2)