+++ doc/whats_new.rst
@@ -14,7 +14,11 @@ New features
 
 
    - Incremental fit for :class:`GaussianNB <naive_bayes.GaussianNB>`.
+
    - Add ``sample_weight`` support to :class:`dummy.DummyClassifier`. By
+     `Arnaud Joly`_.
+
+   - Add the :func:`metrics.label_ranking_average_precision_score` metrics. By
      `Arnaud Joly`_.
 
 
@@ -22,8 +26,12 @@ Enhancements
 ............
 
 
+   - Add support for sample weights in scorer objects.  Metrics with sample
+     weight support will automatically benefit from it.
+
+
 Documentation improvements
-...........................
+..........................
 
 
 Bug fixes
@@ -31,12 +39,17 @@ Bug fixes
 
 API changes summary
 -------------------
-    
+
     - :class:`GridSearchCV <grid_search.GridSearchCV>` and
       :func:`cross_val_score <cross_validation.cross_val_score>` and other
       meta-estimators don't convert pandas DataFrames into arrays any more,
       allowing DataFrame specific operations in custom estimators.
 
+    - :func:`multiclass.fit_ovr`, :func:`multiclass.predict_ovr`,
+      :func:`predict_proba_ovr`,
+      :func:`multiclass.fit_ovo`, :func:`multiclass.predict_ovo`,
+      :func:`multiclass.fit_ecoc` and :func:`multiclass.predict_ecoc`
+      are deprecated. Use the underlying estimators instead.
 
 .. _changes_0_15: